requirements.txt
python-binance
ta
numpy
pandas